audit: bump react major version to 18
This fixes a major vulnerability exposed in a prior version.
Had to remove references to codemirror libraries, which were out of date
and dependent on an old React API.
However this code didn't seem to actually be used.
In the interrim the codemirror has been replaced with an input element.